+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Improvements
 
+- Add function `apply_vectorized` to apply a vectorized function on a geometry column
+  (#588)
 - Add support for renaming column with only difference in casing (#549)
 - Add support to pass a single string for all `column` type parameters if a single
   column should be retained (#523)